+++ cli/cli/Commands/Services/ServicesDeployCommand.cs
@@ -1,6 +1,9 @@
 ﻿using Beamable.Common;
+using Beamable.Common.Api;
+using Beamable.Common.Api.Realms;
 using Beamable.Common.BeamCli;
 using cli.Services;
+using cli.Utils;
 using Newtonsoft.Json;
 using Serilog;
 using Spectre.Console;
@@ -30,6 +33,7 @@ public class ServicesDeployCommand : AppCommand<ServicesDeployCommandArgs>,
 	private BeamoLocalSystem _localBeamo;
 	private BeamoService _remoteBeamo;
 	private ServicesListCommand _servicesListCommand;
+	private IAliasService _aliasService;
 
 	public ServicesDeployCommand(ServicesListCommand servicesListCommand) :
 		base("deploy",
@@ -66,6 +70,7 @@ public class ServicesDeployCommand : AppCommand<ServicesDeployCommandArgs>,
 		_ctx = args.AppContext;
 		_localBeamo = args.BeamoLocalSystem;
 		_remoteBeamo = args.BeamoService;
+		_aliasService = args.AliasService;
 
 		var isDockerRunning = await _localBeamo.CheckIsRunning();
 		if (!isDockerRunning)
@@ -73,6 +78,22 @@ public class ServicesDeployCommand : AppCommand<ServicesDeployCommandArgs>,
 			throw new CliException(
 				"Docker is not running in this machine. Please start Docker before running this command.",
 				Beamable.Common.Constants.Features.Services.CMD_RESULT_CODE_DOCKER_NOT_RUNNING, true);
+		}
+
+		var cid = _ctx.Cid;
+		if (!AliasHelper.IsCid(cid))
+		{
+			try
+			{
+				var aliasResolve = await _aliasService.Resolve(cid).ShowLoading("Resolving alias...");
+				cid = aliasResolve.Cid.GetOrElse(() => throw new CliException("Invalid alias"));
+				_ctx.Set(cid, _ctx.Pid, _ctx.Host);
+			}
+			catch (Exception)
+			{
+				AnsiConsole.WriteLine($"Unable to resolve alias for '{cid}'");
+				return;
+			}
 		}
 
 		try